+++ app/(api)/_datalib/_services/Users.ts
@@ -1,10 +1,13 @@
 import revalidateCache from '@actions/revalidateCache';
 import prisma from '../_prisma/client';
 import { UserInput } from '@datatypes/User';
+import { ApolloContext } from '@datalib/apolloServer';
 
 export default class Users {
   // CREATE
-  static async create(input: UserInput) {
+  static async create(input: UserInput, ctx: ApolloContext) {
+    if (!ctx.isOwner) return null; // TODO: Possibly some better error message.
+
     const user = await prisma.user.create({
       data: input,
     });
@@ -36,7 +39,9 @@ export default class Users {
   }
 
   // UPDATE
-  static async update(id: string, input: UserInput) {
+  static async update(id: string, input: UserInput, ctx: ApolloContext) {
+    if (!ctx.isOwner) return null; // TODO: Possibly some better error message.
+
     try {
       const user = await prisma.user.update({
         where: {
@@ -52,7 +57,9 @@ export default class Users {
   }
 
   // DELETE
-  static async delete(id: string) {
+  static async delete(id: string, ctx: ApolloContext) {
+    if (!ctx.isOwner) return null; // TODO: Possibly some better error message.
+
     try {
       await prisma.user.delete({
         where: {
